Allow group to have named depenencies
With this change, a group will be able to define a list of "needs"
that are names of other groups. When executing, groups will run their
dependencies first. In the event a dependency fails, the group will not
run.

When a "path" fails, if other paths don't depend on the failing group,
then they will continue to run. This will allow as much work as possible
to happen.