Allow group to have named depenencies (#50)
With this change, a group will be able to define a list of "needs"
that are names of other groups. When executing, groups will run their
dependencies first. In the event a dependency fails, the group will not
run.

When a "path" fails, if other paths don't depend on the failing group,
then they will continue to run. This will allow as much work as possible
to happen.

The order that groups will execute will now be stable, and list (both
root and doctor)
let the user know the calculated run order.

Closes https://github.com/ethankhall/scope/issues/45
Closes https://github.com/ethankhall/scope/issues/39
Closes https://github.com/ethankhall/scope/issues/46