Create test for glob walker
This adds two tests, one without an abs path (like would be without
work_dir), and one with an abs path.

There was a bug fixed yesterday around this, but was hard to test.
This adds a new abstraction `FileSystem` that's intent is to abstract
away filesystem access, to allow us to write unit tests without needing
to setup the filesystem in a specific way.

This trait will probably grow as we add more unit tests.