Improve UX by reducing output
The goal of this code is to enable `--progress` as a root logging option.
By default, if `scope` is run within a `tty` then the only real output
should be progress bar updates.

This change also removes many of the logging options to support only
`-v / --verbose` option. By default we'll log much less.

With this, we're able to write to stdout without needing to go through
tracing, removing some of the formatting requirements.

With this change, we're adding support for `report_stdout!` which will
allow developers to write to stdout. This returns a future, so all calls
need to be `async`. It will also "log" the line, so the output will show
up in logs.