Allow groups to configure task resolution
DoctorGroup will have an optional field `include` that will determine if
the group should be included by default when determine task to run.

The `include` option has no effect if `--only` is used.

The default behavior is to include the group; however in the case where
there are shared configs, the shared configs will now specify
`include: when-required` and create a "root" group that declares all the
groups that are required for the app.