Improve UX by reducing output (#99)
The goal of this code is to enable `--progress` as a root logging
option. By default, if `scope` is run within a `tty` then the only real
output should be progress bar updates.

This change also removes many of the logging options to support only `-v
/ --verbose` option. By default we'll log much less.

With this, we're able to write to stdout without needing to go through
tracing, removing some of the formatting requirements.

With this change, we're adding support for `report_stdout!` which will
allow developers to write to stdout. This returns a future, so all calls
need to be `async`. It will also "log" the line, so the output will show
up in logs.

## Output

New CLI options

```
Options:
  -v, --verbose...
          A level of verbosity, and can be used multiple times

      --progress <PROGRESS>
          Set the progress output. Use plain to disable updating UI

          [env: SCOPE_OUTPUT_PROGRESS=]
          [default: auto]

          Possible values:
          - auto:  Determine output format based on execution context
          - plain: Standard output, no progress bar, no auto-updating output
          - tty:   Use progress bar

```


![demo1](https://github.com/oscope-dev/scope/assets/794802/a4fcd061-f995-4074-8a45-ec22daa842ec)

![demo2](https://github.com/oscope-dev/scope/assets/794802/76480dcc-175f-419e-acb3-4c063e06d3f6)