Include summary at end of doctor run
When `scope doctor run` finishes, it will print the number of groups
that succeeded, failed, and were skipped. This will allow people to
know that something was done.

This also fixes a bug, where the progress bar would be stuck at the max
length instead of counting up showing progress / percent complete