Support templated commands
This supports adding `{{ working_dir }}` to commands, to create parity
with file checking. It also allows us to build more template support
in the future should we need/want to support it.

Additionally this update makes it so that integ tests will import "test dirs",
allowing the tests to have purpose built config for the single test
instead of relying on config and filtering like was done before.

This allows for a more "normal" experince of testing, and validation like
a user would expect.