Load env vars from a shared config location (#106)
This PR makes it possible for scope to load environment variables from a
"shared" .env file. This is helpful for scope developers that need to
deploy credentials (i.e. GitHub App or GitHub tokens) to app engineers.

## Testing
With a clean env and a `../etc/scope.env` file populated with a
`GH_TOKEN`, I was able to submit a report:
```
❯ bin/scope report -- echo foobar
 INFO Report was uploaded to https://github.com/Gusto/gusto_scope_testing/issues/4.
```

Removing this file caused `scope report` to fail with the expected
configuration error:
```
❯ bin/scope report -- echo foobar
 WARN Unable to upload to github: GH_TOKEN env var was not set with token to access GitHub
```