Refactor report generation for group report extras
With this change, a group is not able to provide a new field
`reportExtraDetails`, that contains a map of `string:string`, this map
will run _after_ the group has executed regardless if it succeeded or
failed.

The output will be placed into a table in the markdown report.

This allows groups to add metadata about their targets to reports, making
the reports more useful.

This change includes some refactoring work (mainly around naming) that
will come in a later PR. There is also some refactoring to be done around
de-duplicating data, but wanted to exclude that for now as this PR is
already large.