Use templates for report rendering (#124)
This PR refactors the report rendering implementation to use jinja
templates. This paves the way for per-location or configurable
templates.

Tests of report rendering were also added. There is a lot of new code
added here, and some feels duplicative. I'd like to follow up this PR
with a refactoring one that massages things a little more.

Manual test output to follow in comments.