Improve telemetry output
1. Actions and groups are now marked as errored when they fail
2. Set the otel.name so it includes the group or action name

This lets us easily query/aggregate which groups or actions are failing most often.

This fixes #185 and improves on the solution for #154 implemented in #157.

We accomplish the dynamic span naming by leveraging one of the "special fields" detailed here
https://docs.rs/tracing-opentelemetry/latest/tracing_opentelemetry/

The naming convention is loosely based on the OTEL http semantic convention of "{method} {url}".
Several other of the semantic convention standards are similar as well.
https://opentelemetry.io/docs/specs/semconv/http/http-spans/

In order to access the OTEL Span's `set_status()` method, I needed to upgrade all of our opentelemetry dependencies to the latest version.
When I did this, traces were no longer being properly flushed before exiting the application,
so I updated the implementation based off of this example from the opentelemetry-otlp crate.
https://github.com/tokio-rs/tracing-opentelemetry/blob/v0.1.x/examples/opentelemetry-otlp.rs