Add verify: false option to doctor group action fixes.
Sometimes you may want to trigger a fix based on a file change
as well as some other condition, like the value of an environment variable,
but the fix does not change the condition.

Or, perhaps there isn't a good (or fast maybe) way to check that things are ok,
so you want the fix to run whenever `--no-cache` is passed.

Currently, the check always runs after the fix to verify that the fix worked.
This commit introduces a way to skip running the check a second time as verification.
This functionality should only really be used as a last resort option,
but it should be possible for the small set of edge cases where it's needed.