Fixes telemetry via http (#191)
Between opentelemetry-otlp v0.15 and 0.16 a breaking change landed.
-
https://github.com/pitoniak32/opentelemetry-rust/blob/main/opentelemetry-otlp/CHANGELOG.md#v0160
- https://github.com/open-telemetry/opentelemetry-rust/pull/1706

So when we upgraded to 0.27 (#186),
we inadvertantly broke people using `SCOPE_OTEL_PROTOCOL=http`. It did
not affect people using `grpc`.

To maintain consistency with grpc (and compatibility with version prior
to PR #186), we append `v1/traces` to the `SCOPE_OTEL_ENDPOINT` when the
protocol is set to `http`.