Updates to get release pipeline working again (#202)
This build errored out

https://github.com/oscope-dev/scope/actions/runs/15145723045/job/42580832622
with this error message
> This is a scheduled Ubuntu 20.04 retirement. Ubuntu 20.04 LTS runner
will be removed on 2025-04-15. For more details, see
https://github.com/actions/runner-images/issues/11101

Updating the runners accordingly.
Using latest so we don't end up on unsupported runners in the future.

This required us to specify the runner versions explicitly.
See this `cargo-dist` issue for details:
https://github.com/axodotdev/cargo-dist/issues/1760

After that, the pipeline was getting deadlocks.
https://github.com/github/vscode-github-actions/issues/135