Fixes caching bug
PR #203 changed the way we handle checks with both paths and commands however,
we weren't properly updating the cache when the path was invalid but the command succeeded.
This resulted in the paths never being cached for checks that have both paths and commands.

This is now fixed by ensuring we always update the cache when the fix is not needed.
This does result in us updating the cache every time for valid caches, but there's no noticibly difference in performance.