Cache at action level instead of group (#213)
#208 fixed one caching bug by changing when we update the cache.  
I'm unsure if that PR introduced this bug (it may have) but this PR
fixes another.

When ever the _same_ file path is used to cache on in subsequent
actions, the cache is updated when the first succeeds, so the cache is
read as valid for the second action when it should be invalid.

Example config file

```yaml
apiVersion: scope.github.com/v1alpha
kind: ScopeDoctorGroup
metadata:
  name: ssh
  description: Setup SSH keys for GitHub authentication
spec:
  include: when-required
  needs: []
  actions:
    - name: private-key-exists
      description: Ensures SSH private key exists
      check:
        paths:
          - ~/.ssh/id_ed25519
        commands:
          - ./bin/ssh.sh check private-key
      fix:
        commands:
          - ./bin/ssh.sh fix private-key
    - name: public-key-exists
      description: Ensures SSH public key exists
      check:
        paths:
          - ~/.ssh/id_ed25519.pub
        commands:
          - ./bin/ssh.sh check public-key
      fix:
        commands:
          - ./bin/ssh.sh fix public-key
    - name: key-added
      description: Uploads the pub key to GitHub
      check:
        # The paths in this were previously erroneously cached too early
        # so the check was always passing
        paths:
          # if either of these files have changed, we need to re-check
          - ~/.ssh/id_ed25519
          - ~/.ssh/id_ed25519.pub
        commands:
          - ./bin/ssh.sh check github
      fix:
        commands:
          - ./bin/ssh.sh fix github
```


This because we were keying _only_ off the group name and the fully
qualified file name.

This PR updates the cache to _also_ key off of the action name, so that
each check has a unique key in the cache.

Old cache format

```json
{
  "checksums": {
    "rails-group": {
      "/path/to/file1.txt": "abc123",
      "/path/to/file2.txt": "def456"
    },
    "ruby-group": {
      "/path/to/ruby-version": "xyz789"
    }
  }
}
```

New cache format

```json
{
  "checksums": {
    "rails-group": {
        "action-1": { "/path/to/file1.txt": "abc123" },
        "action-2": { "/path/to/file2.txt": "def456" }
    },
    "ruby-group": {
      "action-3": { "/path/to/ruby-version": "xyz789" }
    }
  }
}
```

_IMPORTANT:_ This will invalidate any existing cache on a user's
machine. The first time a user runs with this new version, we will
invalidate all of their existing caches.