Add ability to skip groups (#215)
We've found it desirable to be able to skip entire groups, if certain
conditions are met.
Previously, we did this by ensuring all checks in all actions in a group
always exited with success, which works, but is inconvenient, a little
error prone, and does (minimally) increase execution time because we
still have to execute all the checks for all the actions. It's also
unclear to the user that we are skipping those groups.

This PR adds the ability to bypass a group _either_ by setting
`spec.skip: true` _or_ by adding a `spec.skip.command`.
When the `command` returns `0` (success), the group is skipped.  
When it fails (returns non-0), the group is executed.  

When `spec.skip` is omitted, we default to `false`, which makes this a
backward compatible change.

Examples
---

### Skip Omitted

Omitting the new field is equivalent to `spec.skip: false` and will
behave the same as it does today (i.e. the group is run).

```yaml
apiVersion: scope.github.com/v1alpha
kind: ScopeDoctorGroup
metadata:
  name: group-skip-default
  description: Test group with skip field omitted (defaults to false)
spec:
  actions:
    - name: should-run
      check:
        commands:
          - echo "This check should run"
      fix:
        commands:
          - echo "This fix should run"
```

### Skip False

When `skip: false`, the behavior is the same as it is today (i.e. the
group is run).

```yaml
apiVersion: scope.github.com/v1alpha
kind: ScopeDoctorGroup
metadata:
  name: group-skip-boolean-false
  description: Test group with skip boolean set to false
spec:
  skip: false
  actions:
    - name: should-run
      check:
        commands:
          - echo "This check should run"
      fix:
        commands:
          - echo "This fix should run"
```

### Skip True

When `skip: true`, the group will NOT be executed.   
This can be useful for temporarily disabling a group that is included by
other groups or for putting a group into the correct place in the
dependency tree prior to enabling it.

```yaml
apiVersion: scope.github.com/v1alpha
kind: ScopeDoctorGroup
metadata:
  name: group-skip-boolean
  description: Test group with skip boolean set to true
spec:
  skip: true
  actions:
    - name: should-not-run
      check:
        commands:
          - echo "This check should not run"
      fix:
        commands:
          - echo "This fix should not run"
```

### Skip Command

This is useful for creating feature flags _OR_ allowing users to opt out
of groups.
When the `command` returns `0`, the entire group is skipped, otherwise
the group is run.

Below is an example of using an environment variable to conditionally
skip the group.

```yaml
apiVersion: scope.github.com/v1alpha
kind: ScopeDoctorGroup
metadata:
  name: group-skip-command
  description: Test group with skip command that returns success
spec:
  skip: 
    command: bash -c '[ -n "$APPLESAUCE" ] && [ "$APPLESAUCE" = "1" ]'
  actions:
    - name: should-not-run
      check:
        commands:
          - echo "This check should not run"
      fix:
        commands:
          - echo "This fix should not run"
```