Upgrade toolchain to latest stable
std::env::home_dir() is deprecated in v1.82 because it was broken on Windows,
but instead of removing it, it was fixed in v1.85.

Moving us to the latest stable toolchain (v1.88 at time of writing).

This did require fixing a couple of lints
- https://rust-lang.github.io/rust-clippy/master/index.html#uninlined_format_args
- https://rust-lang.github.io/rust-clippy/master/index.html#useless_vec

The useless_vec fix wasn't strictly _required_ but it came along for the ride
when I ran `clippy cargo --fix`