Use XDG_CACHE_HOME for cache instead of /tmp/scope (#221)
Previously, we were putting the cache file in `/tmp/scope` by default,
which may be convenient for development, but is a not awesome user
experience because the cache gets wiped on reboot. It's a better user
experience to put the cache in the user's home directory so it persists
through a reboot.

We wish to follow the [XDG Base Directory
specification](https://specifications.freedesktop.org/basedir-spec/latest/),
which required that we remove the dependency on the `directories` crate.
See https://github.com/dirs-dev/directories-rs/issues/47 for more
information, but the TL;DR is the maintainer of that crate (also the
maintainer of the `dirs` crate) refuses to support XDG on MacOS.
I couldn't find a suitable replacement and we only currently support
Unix environments anyway, so it was relatively simple to do so.
If we need to support Windows in the future, it wouldn't be terribly
hard to add Windows support by writing to `APPDATA` and `LOCALAPPDATA`.

In order to not bust the user's cache, if we find a
`/tmp/scope/cache-file.json` file _and_ the user didn't explicitly
specify `/tmp/scope` as the cache directory, we move their existing
cache file into the new location for them.

Test coverage was expanded prior to making this change as well.